requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk>=3.0.0
anthropic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
